+++ consensus/block.go
@@ -8,7 +8,7 @@ import (
 	"github.com/pokt-network/pocket/shared/codec"
 )
 
-func (m *ConsensusModule) commitBlock(block *typesCons.Block) error {
+func (m *consensusModule) commitBlock(block *typesCons.Block) error {
 	m.nodeLog(typesCons.CommittingBlock(m.Height, len(block.Transactions)))
 
 	// Store the block in the KV store
@@ -43,7 +43,7 @@ func (m *ConsensusModule) commitBlock(block *typesCons.Block) error {
 
 // IMPROVE(olshansky): ensure these persistence operations are atomic, we can't commit block without
 // committing the transactions and metadata at the same time
-func (m *ConsensusModule) storeBlock(block *typesCons.Block, blockProtoBytes []byte) error {
+func (m *consensusModule) storeBlock(block *typesCons.Block, blockProtoBytes []byte) error {
 	store := m.utilityContext.GetPersistenceContext()
 	// Store in KV Store
 	if err := store.StoreBlock(blockProtoBytes); err != nil {
@@ -66,7 +66,7 @@ func (m *ConsensusModule) storeBlock(block *typesCons.Block, blockProtoBytes []b
 
 // TODO: Add unit tests specific to block validation
 // IMPROVE: (olshansky) rename to provide clarity of operation. ValidateBasic() is typically a stateless check not stateful
-func (m *ConsensusModule) validateBlockBasic(block *typesCons.Block) error {
+func (m *consensusModule) validateBlockBasic(block *typesCons.Block) error {
 	if block == nil && m.Step != NewRound {
 		return typesCons.ErrNilBlock
 	}
@@ -75,8 +75,8 @@ func (m *ConsensusModule) validateBlockBasic(block *typesCons.Block) error {
 		return typesCons.ErrBlockExists
 	}
 
-	if block != nil && unsafe.Sizeof(*block) > uintptr(m.consGenesis.MaxBlockBytes) {
-		return typesCons.ErrInvalidBlockSize(uint64(unsafe.Sizeof(*block)), m.consGenesis.MaxBlockBytes)
+	if block != nil && unsafe.Sizeof(*block) > uintptr(m.consGenesis.GetMaxBlockBytes()) {
+		return typesCons.ErrInvalidBlockSize(uint64(unsafe.Sizeof(*block)), m.consGenesis.GetMaxBlockBytes())
 	}
 
 	// If the current block being processed (i.e. voted on) by consensus is non nil, we need to make
@@ -93,7 +93,7 @@ func (m *ConsensusModule) validateBlockBasic(block *typesCons.Block) error {
 }
 
 // Creates a new Utility context and clears/nullifies any previous contexts if they exist
-func (m *ConsensusModule) refreshUtilityContext() error {
+func (m *consensusModule) refreshUtilityContext() error {
 	// Catch-all structure to release the previous utility context if it wasn't properly cleaned up.
 	// Ideally, this should not be called.
 	if m.utilityContext != nil {