+++ persistence/genesis.go
@@ -12,7 +12,7 @@ import (
 // TODO(andrew): generalize with the `actors interface`
 
 // WARNING: This function crashes the process if there is an error populating the genesis state.
-func (m *PersistenceModule) populateGenesisState(state *types.PersistenceGenesisState) {
+func (m *persistenceModule) populateGenesisState(state modules.PersistenceGenesisState) {
 	log.Println("Populating genesis state...")
 
 	// REFACTOR: This business logic should probably live in `types/genesis.go`
@@ -155,7 +155,9 @@ func (m *PersistenceModule) populateGenesisState(state *types.PersistenceGenesis
 }
 
 // TODO(pocket/issues/149): All of the functions below following a structure similar to `GetAll<Actor>`
-//  can easily be refactored and condensed into a single function using a generic type or a common
+//
+//	can easily be refactored and condensed into a single function using a generic type or a common
+//
 // interface.
 func (p PostgresContext) GetAllAccounts(height int64) (accs []modules.Account, err error) {
 	ctx, tx, err := p.GetCtxAndTx()
@@ -209,9 +211,9 @@ func (p PostgresContext) GetAllApps(height int64) (apps []modules.Actor, err err
 	if err != nil {
 		return nil, err
 	}
-	var actors []types.BaseActor
+	var actors []*types.Actor
 	for rows.Next() {
-		var actor types.BaseActor
+		var actor *types.Actor
 		actor, height, err = p.GetActorFromRow(rows)
 		if err != nil {
 			return
@@ -224,7 +226,7 @@ func (p PostgresContext) GetAllApps(height int64) (apps []modules.Actor, err err
 		if err != nil {
 			return
 		}
-		apps = append(apps, p.BaseActorToActor(actor, types.ActorType_App))
+		apps = append(apps, actor)
 	}
 	return
 }
@@ -238,9 +240,9 @@ func (p PostgresContext) GetAllValidators(height int64) (vals []modules.Actor, e
 	if err != nil {
 		return nil, err
 	}
-	var actors []types.BaseActor
+	var actors []*types.Actor
 	for rows.Next() {
-		var actor types.BaseActor
+		var actor *types.Actor
 		actor, height, err = p.GetActorFromRow(rows)
 		if err != nil {
 			return
@@ -253,7 +255,7 @@ func (p PostgresContext) GetAllValidators(height int64) (vals []modules.Actor, e
 		if err != nil {
 			return
 		}
-		vals = append(vals, p.BaseActorToActor(actor, types.ActorType_Val))
+		vals = append(vals, actor)
 	}
 	return
 }
@@ -267,9 +269,9 @@ func (p PostgresContext) GetAllServiceNodes(height int64) (sn []modules.Actor, e
 	if err != nil {
 		return nil, err
 	}
-	var actors []types.BaseActor
+	var actors []*types.Actor
 	for rows.Next() {
-		var actor types.BaseActor
+		var actor *types.Actor
 		actor, height, err = p.GetActorFromRow(rows)
 		if err != nil {
 			return
@@ -282,7 +284,7 @@ func (p PostgresContext) GetAllServiceNodes(height int64) (sn []modules.Actor, e
 		if err != nil {
 			return
 		}
-		sn = append(sn, p.BaseActorToActor(actor, types.ActorType_Node))
+		sn = append(sn, actor)
 	}
 	return
 }
@@ -296,9 +298,9 @@ func (p PostgresContext) GetAllFishermen(height int64) (f []modules.Actor, err e
 	if err != nil {
 		return nil, err
 	}
-	var actors []types.BaseActor
+	var actors []*types.Actor
 	for rows.Next() {
-		var actor types.BaseActor
+		var actor *types.Actor
 		actor, height, err = p.GetActorFromRow(rows)
 		if err != nil {
 			return
@@ -311,22 +313,7 @@ func (p PostgresContext) GetAllFishermen(height int64) (f []modules.Actor, err e
 		if err != nil {
 			return
 		}
-		f = append(f, p.BaseActorToActor(actor, types.ActorType_Fish))
+		f = append(f, actor)
 	}
 	return
-}
-
-// TODO (Team) deprecate with interface #163 <Bumped to #149> as #163 is getting large
-func (p PostgresContext) BaseActorToActor(ba types.BaseActor, actorType types.ActorType) *types.Actor {
-	actor := new(types.Actor)
-	actor.ActorType = actorType
-	actor.Address = ba.Address
-	actor.PublicKey = ba.PublicKey
-	actor.StakedAmount = ba.StakedTokens
-	actor.GenericParam = ba.ActorSpecificParam
-	actor.PausedHeight = ba.PausedHeight
-	actor.UnstakingHeight = ba.UnstakingHeight
-	actor.Output = ba.OutputAddress
-	actor.Chains = ba.Chains
-	return actor
 }