+++ shared/modules/module.go
@@ -1,5 +1,10 @@
 package modules
 
+import (
+	"github.com/pokt-network/pocket/shared/crypto"
+	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
+)
+
 type Module interface {
 	InitializableModule
 	IntegratableModule
@@ -18,6 +23,21 @@ type InterruptableModule interface {
 
 type InitializableModule interface {
 	GetModuleName() string
-	InitConfig(pathToConfigJSON string) (IConfig, error)
-	InitGenesis(pathToGenesisJSON string) (IGenesis, error)
+	Create(runtime RuntimeMgr) (Module, error)
+}
+
+type ConfigurableModule interface {
+	ValidateConfig(Config) error
+}
+
+type GenesisDependentModule interface {
+	ValidateGenesis(GenesisState) error
+}
+
+type KeyholderModule interface {
+	GetPrivateKey() (crypto.PrivateKey, error)
+}
+
+type P2PAddressableModule interface {
+	GetP2PAddress() cryptoPocket.Address
 }