+++ shared/modules/persistence_module.go
@@ -9,6 +9,9 @@ import (
 
 type PersistenceModule interface {
 	Module
+	ConfigurableModule
+	GenesisDependentModule
+
 	NewRWContext(height int64) (PersistenceRWContext, error)
 	NewReadContext(height int64) (PersistenceReadContext, error)
 	GetBlockStore() kvstore.KVStore