+++ utility/test/account_test.go
@@ -6,12 +6,11 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/pokt-network/pocket/shared/modules"
-	"github.com/pokt-network/pocket/shared/test_artifacts"
-	"github.com/pokt-network/pocket/utility/types"
-
+	"github.com/pokt-network/pocket/runtime/test_artifacts"
 	"github.com/pokt-network/pocket/shared/crypto"
+	"github.com/pokt-network/pocket/shared/modules"
 	"github.com/pokt-network/pocket/utility"
+	"github.com/pokt-network/pocket/utility/types"
 	"github.com/stretchr/testify/require"
 )