+++ consensus/CHANGELOG.md
@@ -11,6 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Stores transactions alongside blocks during `commit`
 - Added current block `[]TxResult` to the module
 
+### [#235](https://github.com/pokt-network/pocket/pull/235) Config and genesis handling
+
+- Updated to use `RuntimeMgr`
+- Made `ConsensusModule` struct unexported
+- Updated tests and mocks
+- Removed some cross-module dependencies
+
 ## [0.0.0.5] - 2022-10-06
 
 - Don't ignore the exit code of `m.Run()` in the unit tests